backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
APScheduler==3.10.4
praw==7.8.1
python-dotenv==1.0.1
